Fix colours overflow and create a shape builder

+++ src/graphics/colours.rs
@@ -5,10 +5,10 @@
 /// Colour with red, green and blue values + an alpha channel.
 #[derive(Debug, Copy, Clone, Eq, PartialEq)]
 pub struct RGBA {
-    pub r: u8,
-    pub g: u8,
-    pub b: u8,
-    pub a: u8,
+    pub r: u32,
+    pub g: u32,
+    pub b: u32,
+    pub a: u32,
 }
 
 impl RGBA {
@@ -17,7 +17,7 @@ impl RGBA {
     /// `r`, `g`, `b` and `a` are in range 0 to 255, where 0 is the non-colour 
     /// (black) and 255 the full colour (red, green or blue). For `a` being the 
     /// alpha channel, 0 transparent and 255 is completely visible.
-    pub fn new(r: u8, g: u8, b: u8, a: u8) -> Self {
+    pub fn new(r: u32, g: u32, b: u32, a: u32) -> Self {
         Self { r, g, b, a, }
     }
 
@@ -25,7 +25,7 @@ impl RGBA {
     /// channel.
     /// 
     /// Converts the hexadecimal `value` into a "RGBA" structure.
-    pub fn from_hex(value: u8) -> Self {
+    pub fn from_hex(value: u32) -> Self {
         Self {
             r: ((value >> 24) & 0xFF) / 255,    // RR byte
             g: ((value >> 16) & 0xFF) / 255,    // GG byte
@@ -35,7 +35,7 @@ impl RGBA {
     }
 
     /// Converts the `RGBA` values into an hexadecimal value.
-    pub fn to_hex(&self) -> u8 {
+    pub fn to_hex(&self) -> u32 {
         ((self.r & 0xFF) << 24) 
             + ((self.g & 0xFF) << 16) 
             + ((self.b & 0xFF) << 8)
